requirements.txt
langchain
langchain-openai
python-dotenv
pytest
coverage
langchain-community
faiss-cpu